app/src/requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
databases==0.6.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
zipp==3.8.1
